chore: rename pages
