+++ src/App.js
@@ -1,19 +1,18 @@
-import React from 'react';
-import './App.css';
+import React from "react";
+import "./App.css";
 import { Route, Switch } from "react-router-dom";
-import Main from "./pages/mainpage/mainpage";
-import Valid from "./pages/validpage/validpage";
-import Vote from "./pages/votepage/votepage";
-
+import Main from "./pages/HomePage";
+import Valid from "./pages/ValidateVoterPage";
+import Vote from "./pages/VotePage";
 
 function App() {
   return (
-      <Switch>
-        <Route exact path="/" component={Main} />
-        <Route exact path="/validate" component={Valid} />
-        <Route exact path="/vote" component={Vote} />
-      </Switch>
-  )
+    <Switch>
+      <Route exact path="/" component={Main} />
+      <Route exact path="/validate" component={Valid} />
+      <Route exact path="/vote" component={Vote} />
+    </Switch>
+  );
 }
 
 export default App;
